refactor: add validation to appointment and medical event DTOs and update file structure
