+++ src/medical-scheduling/appointments/appointments.controller.ts
@@ -8,7 +8,7 @@ import {
   Query,
 } from '@nestjs/common';
 import { AppointmentsService } from './appointments.service';
-import { CreateAppointmentDto } from './dto/create-appointments.dto';
+import { CreateAppointmentDto } from './dto/create-appointment.dto';
 import { status_type } from '@prisma/client';
 
 @Controller('appointments')
