+++ src/medical-scheduling/medical-events/dto/create-medical-event.dto.ts
@@ -1,12 +1,39 @@
+import { IsNotEmpty, IsOptional, IsString } from 'class-validator';
+
 export class CreateMedicalEventDto {
+  @IsNotEmpty()
+  @IsString()
   appointment_id: string;
+
+  @IsNotEmpty()
+  @IsString()
   patient_id: string;
+
+  @IsNotEmpty()
+  @IsString()
   physician_id: string;
-  physician_comments: string;
+
+  @IsOptional()
+  @IsString()
+  physician_comments?: string;
+
+  @IsNotEmpty()
+  @IsString()
   main_diagnostic_cie: string;
-  consultation_reason: string;
-  evolution: string;
-  procedure: string;
-  treatment: string;
+
+  @IsOptional()
+  @IsString()
+  evolution?: string;
+
+  @IsOptional()
+  @IsString()
+  procedure?: string;
+
+  @IsOptional()
+  @IsString()
+  treatment?: string;
+
+  @IsNotEmpty()
+  @IsString()
   tenant_id: string;
 }
